Re-arrange module structure
So things that are about betting are in betting and not everything is
under "party".
