+++ src/betting/joint_output.rs
