+++ src/betting/randomize.rs
